File Upload and Download
Testing ability to download a file from a web page that is referred to by a variable name. Currently pulls image from local folder on the computer and them opens page to download it. Will replace this with the ability to save the variable as the outcome from OpenSCAD in the future.